added layerd decomposition model, full pass is possible, bugs persist

+++ InputProcessing/frameIterator.py
@@ -20,7 +20,7 @@ class FrameIterator(object):
         self.device = device
 
     def __getitem__(self, idx):
-        img = torch.from_numpy(cv2.resize(np.float64(cv2.imread(self.images[idx])), self.frame_size)).permute(2, 0, 1) / 255.
+        img = torch.from_numpy(cv2.resize(np.float32(cv2.imread(self.images[idx])), self.frame_size)).permute(2, 0, 1) / 255.
         img = F.normalize(img, [0.485, 0.456, 0.406], [0.229, 0.224, 0.225])
         
         return img.to(self.device)
